requirements.txt
asgiref==3.11.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
sqlparse==0.5.4
tzdata==2025.2
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
